Make builtins header, reorganize stuff

+++ utils/timer.hpp
@@ -6,6 +6,6 @@ void set_timer() {
   timer = std::chrono::high_resolution_clock::now();
 }
 
-ll elapsed() {
+long long elapsed() {
   return (chrono::duration_cast<chrono::milliseconds>(chrono::high_resolution_clock::now() - timer)).count();
 }
